# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
...
paths:
...
  /api/mcp:
    post:
      operationId: mcpPost
      summary: MCP Streamable HTTP (JSON-RPC)
      description: |
        Methods: initialize, tools/list, tools/call, notifications/initialized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: JSON-RPC response
    get:
      operationId: mcpGet
      summary: MCP SSE stream
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: {}
    delete:
      operationId: mcpDelete
      summary: Terminate MCP session
      responses:
        "200":
          description: Terminated
  /api/mcp/tools:
    get:
      operationId: listMcpTools
      summary: List MCP tools
      responses:
        "200":
          description: Tools list
    post:
      operationId: executeMcpTool
      summary: Execute an MCP tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                args:
                  type: object
      responses:
        "200":
          description: Tool result
    patch:
      operationId: updateMcpToolsConfig
      summary: Update MCP tool configuration
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated config
...
  /api/workflows:
    get:
      operationId: listWorkflows
      summary: List all workflow YAML definitions from resources/flows/
      responses:
        "200":
          description: Workflow list
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
                        version:
                          type: string
                        trigger:
                          type: object
                        steps:
                          type: array
                        yamlContent:
                          type: string
    post:
      operationId: createWorkflow
      summary: Create a new workflow YAML file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, yamlContent]
              properties:
                id:
                  type: string
                yamlContent:
                  type: string
      responses:
        "201":
          description: Workflow created
        "400":
          description: Invalid input
        "409":
          description: Workflow already exists
  /api/workflows/{id}:
    get:
      operationId: getWorkflow
      summary: Get a specific workflow by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow detail
        "404":
          description: Workflow not found
    put:
      operationId: updateWorkflow
      summary: Update a workflow YAML file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [yamlContent]
              properties:
                yamlContent:
                  type: string
      responses:
        "200":
          description: Workflow updated
        "400":
          description: Invalid input
        "404":
          description: Workflow not found
    delete:
      operationId: deleteWorkflow
      summary: Delete a workflow YAML file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow deleted
        "404":
          description: Workflow not found
  /api/workflows/{id}/trigger:
    post:
      operationId: triggerWorkflow
      summary: Trigger a workflow run inside a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId]
              properties:
                workspaceId:
                  type: string
                triggerPayload:
                  type: string
      responses:
        "201":
          description: Workflow run created
          content:
            application/json:
              schema:
                type: object
                required: [workflowRunId, taskIds]
                properties:
                  workflowRunId:
                    type: string
                  taskIds:
                    type: array
                    items:
                      type: string
        "400":
          description: Invalid input
        "404":
          description: Workflow or workspace not found

  # ── Debug (Rust desktop only) ──
...